business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SetTransmitFrequency.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SetTransmitFrequency implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+l ");
        messageString.append(command.getParameter("TX Frequency"));
        messageString.append("\n");
        return messageString;
    }
}
